backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20
kokoro>=0.9.4
soundfile==0.12.1
groq==0.13.0
python-dotenv==1.0.1
